add a couple of useful functions and few minor changes

+++ src/types.h
@@ -68,8 +68,8 @@ namespace ymir {
     *
     * \brief Type of stored probabilities of different events.
     */
-    typedef double prob_t;
-//    typedef long double prob_t;
+//    typedef double prob_t;
+    typedef long double prob_t;
     #endif
